'Changes' - package.json - views (All folder rename)
