+++ src/app/about/page.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ПочемумыAboutIndex from "../../views/Почему-мы";
+import ПочемумыAboutIndex from "../../views/About";
 
 const ПочемумыAboutPage = () => <ПочемумыAboutIndex />;
